[MRG] Pretrained model for topic classification (#494)
* refactor sentiment_analysis to accommodate topic_classifier

* add interface and create example for pretrained topic classifier

* test for topic classifier

* create mkdocs for topic classifier

* minor correction in mkdocs

* Refactor TextClassifier class

* resolve codacy issues

* broken tests fix

* remove trailing whitespaces

* moving constants related to text model's ids into constants.py
