+++ examples/pretrained_models/text_models/sentiment_analysis_example.py
@@ -1,4 +1,4 @@
-from autokeras.pretrained.sentiment_analysis import SentimentAnalysis
+from autokeras.pretrained.text_classifier import SentimentAnalysis
 
 sentiment_cls = SentimentAnalysis()
